release -> to main - 23.12.2024

+++ src/utils/productsHelper.ts
@@ -10,7 +10,8 @@ import {
   updateDoc,
   doc,
   addDoc,
-  deleteDoc
+  deleteDoc,
+  Timestamp
 } from 'firebase/firestore'
 
 
@@ -58,14 +59,7 @@ export const getNewspaperProducts = async (): Promise<ProductType[]> => {
   return products
 }
 
-export const getProductsByCategory = async (categoryName: string): Promise<ProductType[]> => {
-  const categoryCollection = collection(firestore, 'productCategories')
-  const categoryQuery = query(categoryCollection, where('name', '==', categoryName))
-  const categorySnapshot = await getDocs(categoryQuery)
-
-  if (categorySnapshot.docs.length !== 1) throw new Error('Category query - error occur')
-  const categoryId = categorySnapshot.docs[0].id
-
+export const getProductsByCategory = async (categoryId: string): Promise<ProductType[]> => {
   const productsCollection = collection(firestore, 'products')
   const productsQuery = query(productsCollection, where('category', '==', categoryId))
   const productsSnapshot = await getDocs(productsQuery)
@@ -116,18 +110,22 @@ export const getProductsById = async (productIds: string[]): Promise<ProductType
 }
 
 export const upsertProduct = async (product: ProductType) => {
-  const { id: _, ...values } = product
+  const { id, addedToCartDate, ...values } = product
   if (!!!product.id) {
-    await addDoc(collection(firestore, 'products'), { ...values })
+    await addDoc(collection(firestore, 'products'), { ...values, createDate: Timestamp.now() })
   } else {
     await updateDoc(doc(firestore, 'products', product.id), {
-      ...values
+      ...values,
+      updateDate: Timestamp.now()
     })
   }
 }
 
 export const deleteProduct = async (product: ProductType) => {
-  const categoryImage = ref(storage, product.img)
-  await deleteObject(categoryImage)
+  if (!!product.img) {
+    const categoryImage = ref(storage, product.img)
+    await deleteObject(categoryImage)
+  }
+  
   await deleteDoc(doc(firestore, 'products', product.id))
 }
